+++ src/__tests__/GradeBands.ts
@@ -1,0 +1,25 @@
+// import { GradeBands } from '../GradeBands'
+// import { GradeScales } from '../GradeScale'
+
+test.todo('todo')
+// describe('GradeBands', () => {
+//   test('returns grade band based on grade and grade scale type', () => {
+//     expect(getGradeBand('v0', GradeScales.VSCALE)).toEqual(GradeBands.BEGINNER)
+//     expect(getGradeBand('v2', GradeScales.VSCALE)).toEqual(GradeBands.INTERMEDIATE)
+//     expect(getGradeBand('v5', GradeScales.VSCALE)).toEqual(GradeBands.ADVANCED)
+//     expect(getGradeBand('v7', GradeScales.VSCALE)).toEqual(GradeBands.EXPERT)
+
+//     expect(getGradeBand('5.9', GradeScales.YDS)).toEqual(GradeBands.INTERMEDIATE)
+//     expect(getGradeBand('7a', GradeScales.FONT)).toEqual(GradeBands.EXPERT)
+//     expect(getGradeBand('7a', GradeScales.FRENCH)).toEqual(GradeBands.ADVANCED)
+//   })
+//   test('returns unknown gradeband if grade and grade scale type are mismatched', () => {
+//     jest.spyOn(console, 'warn').mockImplementation()
+//     expect(getGradeBand('5.9', GradeScales.VSCALE)).toEqual(GradeBands.UNKNOWN)
+//     expect(console.warn).toHaveBeenCalledWith(
+//       expect.stringContaining(
+//         'Unexpected grade format: 5.9'
+//       )
+//     )
+//   })
+// })
